requirements.txt
mysql-connector-python
pillow
docxtpl
tkinterdnd2
tkcalendar
pandas
openpyxl